ssc-final/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-adk==1.28.1
google-generativeai>=0.8.0
pydantic==2.9.2
python-multipart==0.0.9
python-dotenv==1.0.1
